3619: Rename SmartAttributeEditorMatcher files
