+++ common/src/View/SmartPropertyEditorManager.cpp
@@ -26,9 +26,9 @@
 #include "View/SmartTypeEditorMatcher.h"
 #include "View/SmartColorEditor.h"
 #include "View/SmartDefaultAttributeEditor.h"
-#include "View/SmartAttributeEditorMatcher.h"
 #include "View/SmartFlagsEditor.h"
 #include "View/SmartPropertyEditor.h"
+#include "View/SmartPropertyEditorMatcher.h"
 
 #include <kdl/memory_utils.h>
